Restructure lattice/ package: move Lattice2D, compact, interlaced into submodules.
- supervillain/lattice.py → lattice/two_dimensional.py (Lattice2D, _Lattice2D)
- compact/compact.py → lattice/compact.py (Lattice, Form, d, delta, star, wedge)
- compact/interlaced.py → lattice/interlaced.py
- lattice/__init__.py: shared _dimension, re-exports Lattice2D and _Lattice2D

All existing import sites updated; all 806 tests pass unchanged.
