+++ supervillain/compact/__init__.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 
-from supervillain.compact.compact import Lattice
-from supervillain.compact.compact import Form, d, delta, star, wedge
+from supervillain.lattice.compact import Lattice
+from supervillain.lattice.compact import Form, d, delta, star, wedge
 
 # from supervillain.compact.villain import Villain
